+++ src/contexts/Auth.tsx
@@ -1,14 +1,14 @@
 import { ComponentType, createContext, useContext } from "react";
 import { Navigate, useLocation } from "react-router-dom";
+import { AuthenticatedUser, CognitoGroup } from "types/auth";
 import Icon from "components/Icon";
 import LoaderRing from "components/LoaderRing";
 import { useGetter } from "store/accessors";
-import { AuthenticatedUser, PreRenderCredential } from "slices/auth";
 import { appRoutes } from "constants/routes";
 
 export default function withAuth<Props>(
   Component: ComponentType<Props & { user: AuthenticatedUser }>,
-  requiredCredentials?: PreRenderCredential[]
+  requiredGroups?: CognitoGroup[]
 ) {
   return function Protected(props: Props) {
     const location = useLocation();
@@ -28,9 +28,7 @@ export default function withAuth<Props>(
       );
     }
 
-    if (
-      !(requiredCredentials || []).every((c) => user.credentials.includes(c))
-    ) {
+    if (!(requiredGroups || []).every((g) => user.groups.includes(g))) {
       return (
         <div className="grid content-start place-items-center py-20">
           <Icon type="ExclamationCircleFill" size={80} className="text-red" />
