+++ src/slices/auth.ts
@@ -6,27 +6,7 @@ import {
   getCurrentUser,
   signOut,
 } from "aws-amplify/auth";
-
-type User = null | "loading" | AuthenticatedUser;
-
-/** credentials that can be checked before protected page is rendered*/
-export type PreRenderCredential = "ap";
-/** credential that can only be verified once protected page is rendered
- *  e.g. only after rendering admin:1 can check if user has credential to access admin:1
- */
-export type PostRenderCredential = number; //EndowmentID
-type Credential = PostRenderCredential | PreRenderCredential;
-
-export type AuthenticatedUser = {
-  token: string;
-  credentials: Credential[];
-  email: string;
-  firstName?: string;
-  isSigningOut: boolean;
-};
-
-export const userIsSignedIn = (user: User): user is AuthenticatedUser =>
-  !!(user as AuthenticatedUser).token;
+import { User, userIsSignedIn } from "types/auth";
 
 type State = {
   user: User;
@@ -36,9 +16,7 @@ const initialState: State = {
   user: null,
 };
 
-export const logout = createAsyncThunk("auth/logout", async () => {
-  await signOut();
-});
+export const logout = createAsyncThunk("auth/logout", signOut);
 
 export const loadSession = createAsyncThunk<User, AuthUser | undefined>(
   "auth/loadSession",
@@ -54,19 +32,14 @@ export const loadSession = createAsyncThunk<User, AuthUser | undefined>(
       ]);
 
       //remove standard claims
-      const payload = session.tokens.accessToken.payload;
+      const { endowments = [], "cognito:groups": groups = [] } =
+        session.tokens.accessToken.payload;
       const token = session.tokens.accessToken.toString();
-
-      const credentials = [
-        Array.isArray(payload["cognito:groups"]) &&
-        payload["cognito:groups"].includes("AngelProtocolAdmin")
-          ? ("ap" as Credential)
-          : null,
-      ].filter((item): item is Credential => item !== null);
 
       return {
         token,
-        credentials,
+        groups: groups as string[], //AWS generated so there's a level of safety
+        endowments: endowments as number[],
         /**
          * email is guaranteed as it is the primary verifcation mechanism,
          * and is always retrieved from federated signin
